+++ src/Symfony/Bundle/WebProfilerBundle/Tests/Functional/WebProfilerBundleKernel.php
@@ -14,7 +14,6 @@ namespace Symfony\Bundle\WebProfilerBundle\Tests\Functional;
 use Psr\Log\NullLogger;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
-use Symfony\Bundle\FrameworkBundle\Test\ExceptionSubscriber;
 use Symfony\Bundle\TwigBundle\TwigBundle;
 use Symfony\Bundle\WebProfilerBundle\WebProfilerBundle;
 use Symfony\Component\Config\Loader\LoaderInterface;
@@ -58,11 +57,6 @@ class WebProfilerBundleKernel extends Kernel
             'session' => ['storage_factory_id' => 'session.storage.factory.mock_file'],
             'router' => ['utf8' => true],
         ];
-
-        // If Symfony >= 6.2
-        if (class_exists(ExceptionSubscriber::class)) {
-            $config['catch_all_throwables'] = true;
-        }
 
         $containerBuilder->loadFromExtension('framework', $config);