+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 6.2
 ---
 
- * Add constructor argument `bool $catchThrowable` to `HttpKernel`
+ * Add constructor argument `bool $handleAllThrowable` to `HttpKernel`
  * Add `ControllerEvent::getAttributes()` to handle attributes on controllers
  * Add `#[Cache]` to describe the default HTTP cache headers on controllers
  * Add `absolute_uri` option to surrogate fragment renderers